+++ Core/Data/Context.cs
@@ -4,17 +4,14 @@ namespace AstralKeks.Workbench.Core.Data
 {
     public class Context
     {
-        private readonly string _workspaceDirectory;
-        private readonly string _userspaceDirectory;
-
         public Context(string workspaceDirectory, string userspaceDirectory)
         {
-            _workspaceDirectory = workspaceDirectory ?? throw new ArgumentNullException(nameof(workspaceDirectory));
-            _userspaceDirectory = userspaceDirectory ?? throw new ArgumentNullException(nameof(userspaceDirectory));
+            WorkspaceDirectory = workspaceDirectory ?? throw new ArgumentNullException(nameof(workspaceDirectory));
+            UserspaceDirectory = userspaceDirectory ?? throw new ArgumentNullException(nameof(userspaceDirectory));
         }
 
-        public string WorkspaceDirectory => _workspaceDirectory;
+        public string WorkspaceDirectory { get; }
 
-        public string UserspaceDirectory => _userspaceDirectory;
+        public string UserspaceDirectory { get; }
     }
 }