+++ Core/Data/Resource.cs
@@ -1,8 +1,0 @@
-﻿
-namespace AstralKeks.Workbench.Core.Data
-{
-    internal static class Resource
-    {
-        public const string Namespace = "AstralKeks.Workbench.Core.Resource";
-    }
-}