+++ Core/Management/ResourceManager.cs
@@ -27,7 +27,20 @@ namespace AstralKeks.Workbench.Core.Management
             var resource = new Resource<string>(format, workspaceProvider, userspaceProvider, defaults);
             resource.Read();
         }
-        
+
+        public void InitializeResource(string userspaceDirectory, string directory, string filename)
+        {
+            var userspacePath = GetResourcePath(userspaceDirectory, directory, filename);
+            var embeddedResourceName = GetResourceName(filename);
+
+            var userspaceProvider = new FileResourceProvider(userspacePath);
+            var defaults = new EmbeddedResourceProvider(embeddedResourceName);
+            var format = new StringResourceFormat();
+
+            var resource = new Resource<string>(format, userspaceProvider, defaults);
+            resource.Read();
+        }
+
         public Resource<TObject> GetResource<TObject>(string workspaceDirectory, string userspaceDirectory, 
             string directory, string filename)
         {
@@ -60,9 +73,20 @@ namespace AstralKeks.Workbench.Core.Management
             return _fileSystemManager.GetAbsolutePath(rootDirectory, directory, filename);
         }
 
+        public string[] GetResourcePaths(string rootDirectory, string directory)
+        {
+            return _fileSystemManager.GetFilesInDirectory(rootDirectory, directory);
+        }
+
+        public void RemoveResource(string rootDirectory, string directory, string filename)
+        {
+            var resourcePath = GetResourcePath(rootDirectory, directory, filename);
+            _fileSystemManager.DeleeteFile(resourcePath);
+        }
+
         public string GetResourceName(string name)
         {
-            return $"{Data.Resource.Namespace}.{OperatingSystemManager.Current}.{name}";
+            return $"AstralKeks.Workbench.Core.Resource.{OperatingSystemManager.Current}.{name}";
         }
     }
 }